be/requirements.txt
fastapi
psycopg2-binary
python-dotenv
python-jose[cryptography]
cryptography
sqlalchemy
alembic
uvicorn
google-generativeai
email-validator
pandas
openpyxl
PyPDF2
python-multipart
requests
feedparser